% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_objects.R
\name{as_units}
\alias{as_units}
\title{Coerce to units}
\usage{
as_units(x, unit)
}
\arguments{
\item{x}{Object to coerce.}

\item{unit}{Units or character string (see \code{\link{parse_unit}}).}
}
\description{
Converts an object to \code{units} using \code{\link[units]{as.units}}.
}
\examples{
as_units(1:5)
as_units(1:5, "m/s")
as_units(1:5, parse_unit("m/s"))
dt <- as.difftime(1, units = "hours")
as_units(dt)
as_units(dt, "min")
}
