% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_objects.R
\name{convert_units}
\alias{convert_units}
\title{Convert units}
\usage{
convert_units(x, from = x, to)
}
\arguments{
\item{x}{Object coercible to \code{units}.}

\item{from}{Units or character string (see \code{\link{parse_unit}}).}

\item{to}{Units or character string (see \code{\link{parse_unit}}).}
}
\description{
Converts between different units of measure.
}
\examples{
convert_units(1, "km", "m")
x <- as_units(1, "km")
convert_units(x, to = x)
convert_units(x, to = "m")
}
