% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_objects.R
\name{is_units}
\alias{is_units}
\title{Test if units}
\usage{
is_units(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Tests whether an object is \code{units}.
}
\examples{
is_units(1:5)
is_units(as_units(1:5))
}
