% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_strings.R
\name{parse_unit}
\alias{parse_unit}
\title{Parse unit from string}
\usage{
parse_unit(string)
}
\arguments{
\item{string}{(character) Unit string to parse.}
}
\description{
Converts a character string to \code{units}. If available, unit names are replaced by the corresponding symbol.
}
\details{
The string may only contain unit names, numeric exponents, and operators "/", "*", "^", "(", and ")". Spaces between unit names are interpreted as "*". This includes the following standard formats:
\itemize{
  \item Product power form (e.g. "kg m2 s-1")
  \item Arithmetic expression (e.g. "kg * (m^2 / s)")
  \item ... or a mix of both (e.g. "kg1 (m^2 s-1)")
}

Unlike \code{\link[units]{parse_unit}}:
\itemize{
  \item Input must be a length-one character vector.
  \item Unknown units throw an error (via \code{\link[udunits2]{ud.is.parseable}}).
  \item Unit names are replaced by their symbol (via \code{\link[udunits2]{ud.get.symbol}}). This is a limitation of evaluating the constructed expression against \code{\link[units]{ud_units}}.
  \item An empty string results in \code{units::as.units(1, value = units::unitless)} rather than \code{units::make_unit("")}.
}

For a list of supported units and their notation, see the UDUNITS-2 \href{https://www.unidata.ucar.edu/software/udunits/udunits-2.2.25/doc/udunits/udunits2.html#Database}{Units Database}.
}
\examples{
parse_unit("m/s")
parse_unit("(m)/(s)")
parse_unit("m s-1")
parse_unit("meters/seconds")
parse_unit("meters seconds-1")
parse_unit("kg * (m^2 / s^3)")
parse_unit("kg (m^2 / s^3)")
parse_unit("kg m2 s-3")
parse_unit("in")
parse_unit("in/s")
parse_unit("°")
parse_unit("µm/s")
parse_unit("mm")
parse_unit("H2O+1 m1 s-1")
\dontrun{
parse_unit("unknown")
}
}
